backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0
PyJWT==2.8.0
bcrypt==4.0.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
Pillow==10.0.1
reportlab==4.0.4
openpyxl==3.1.2
pandas==2.1.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2 